pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-chess>=1.999",
    "requests>=2.32.5",
    "websockets>=16.0",
]
